// apps/client/src/lib/api-contract/types.gen.ts
export type LaunchSchema = {
    /**
     * Failed At
     */
    failed_at?: string | null;
    /**
     * Finished At
     */
    finished_at?: string | null;
    /**
     * Id
     */
    id: string;
    /**
     * Is Aborted
     */
    is_aborted?: boolean | null;
    /**
     * Scheduled At
     */
    scheduled_at: string;
    /**
     * Scheduled By
     */
    scheduled_by: string;
    /**
     * Skipped At
     */
    skipped_at?: string | null;
    /**
     * Started At
     */
    started_at?: string | null;
    /**
     * Status
     */
    status: string;
};
// ...
export type ScopedTaskStatus = 'NEW' | 'PENDING' | 'IN_PROGRESS' | 'SUCCESS' | 'FAILED' | 'SKIPPED';

/**
 * TaskListResponse
 */
export type TaskListResponse = {
    /**
     * Tasks
     */
    tasks: Array<TaskSchema>;
};

/**
 * TaskSchema
 */
export type TaskSchema = {
    current_launch?: LaunchSchema | null;
    /**
     * Depends On
     */
    depends_on: Array<string>;
    /**
     * Description
     */
    description: string;
    /**
     * Id
     */
    id: string;
    /**
     * Label
     */
    label: string;
    latest_launch?: LaunchSchema | null;
    /**
     * Spec Id
     */
    spec_id: string;
    status: ScopedTaskStatus;
};