// apps/client/src/lib/api-contract/zod.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

import * as z from 'zod';

/**
 * LaunchSchema
 */
export const zLaunchSchema = z.object({
    failed_at: z.iso.datetime().nullish(),
    finished_at: z.iso.datetime().nullish(),
    id: z.uuid(),
    is_aborted: z.boolean().nullish(),
    scheduled_at: z.iso.datetime(),
    scheduled_by: z.string(),
    skipped_at: z.iso.datetime().nullish(),
    started_at: z.iso.datetime().nullish(),
    status: z.string()
});

/**
 * LogLevel
 */
export const zLogLevel = z.enum([
    'INFO',
    'WARNING',
    'ERROR'
]);

/**
 * LogType
 */
export const zLogType = z.enum(['UNCLASSIFIED', 'FILE']);

/**
 * JournalEntrySchema
 */
export const zJournalEntrySchema = z.object({
    id: z.uuid(),
    level: zLogLevel,
    message: z.string(),
    timestamp: z.iso.datetime(),
    type: zLogType
});

/**
 * JournalResponse
 */
export const zJournalResponse = z.object({
    journal: z.array(zJournalEntrySchema)
});

/**
 * ScopeResponse
 */
export const zScopeResponse = z.object({
    scope_id: z.uuid()
});

/**
 * ScopedTaskStatus
 */
export const zScopedTaskStatus = z.enum([
    'NEW',
    'PENDING',
    'IN_PROGRESS',
    'SUCCESS',
    'FAILED',
    'SKIPPED'
]);

/**
 * TaskSchema
 */
export const zTaskSchema = z.object({
    current_launch: zLaunchSchema.nullish(),
    depends_on: z.array(z.string()),
    description: z.string(),
    id: z.uuid(),
    label: z.string(),
    latest_launch: zLaunchSchema.nullish(),
    spec_id: z.string(),
    status: zScopedTaskStatus
});

/**
 * ScheduleResponse
 */
export const zScheduleResponse = z.object({
    tasks: z.array(zTaskSchema)
});

/**
 * TaskListResponse
 */
export const zTaskListResponse = z.object({
    tasks: z.array(zTaskSchema)
});

/**
 * ValidationError
 */
export const zValidationError = z.object({
    ctx: z.record(z.string(), z.unknown()).optional(),
    input: z.unknown().optional(),
    loc: z.array(z.union([z.string(), z.int()])),
    msg: z.string(),
    type: z.string()
});

/**
 * HTTPValidationError
 */
export const zHttpValidationError = z.object({
    detail: z.array(zValidationError).optional()
});

export const zInitScopeApiScopesScopeIdPostPath = z.object({
    scope_id: z.uuid()
});

/**
 * Successful Response
 */
export const zInitScopeApiScopesScopeIdPostResponse = zScopeResponse;

export const zStopRunApiScopesScopeIdRunDeletePath = z.object({
    scope_id: z.uuid()
});

/**
 * Successful Response
 */
export const zStopRunApiScopesScopeIdRunDeleteResponse = z.void();

export const zGetTasksApiScopesScopeIdTasksGetPath = z.object({
    scope_id: z.uuid()
});

/**
 * Successful Response
 */
export const zGetTasksApiScopesScopeIdTasksGetResponse = zTaskListResponse;

export const zAbortTaskApiScopesScopeIdTasksTaskIdLaunchesLaunchIdDeletePath = z.object({
    scope_id: z.uuid(),
    task_id: z.string(),
    launch_id: z.string()
});

/**
 * Successful Response
 */
export const zAbortTaskApiScopesScopeIdTasksTaskIdLaunchesLaunchIdDeleteResponse = z.void();

export const zGetJournalApiScopesScopeIdTasksTaskIdLaunchesLaunchIdJournalGetPath = z.object({
    scope_id: z.uuid(),
    task_id: z.string(),
    launch_id: z.string()
});

/**
 * Successful Response
 */
export const zGetJournalApiScopesScopeIdTasksTaskIdLaunchesLaunchIdJournalGetResponse = zJournalResponse;

export const zScheduleTaskApiScopesScopeIdTasksTaskIdSchedulePostPath = z.object({
    scope_id: z.uuid(),
    task_id: z.string()
});

/**
 * Successful Response
 */
export const zScheduleTaskApiScopesScopeIdTasksTaskIdSchedulePostResponse = zScheduleResponse;